requirements.txt
aiogram
python-dotenv
pyrogram
tgcrypto
telethon
